Move PublicOnly type into the src directory
